// src/services/generated/models/userModel.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * SEMD API
 * OpenAPI spec version: 1.0.0
 */
import type { UserModelBirthday } from './userModelBirthday';
import type { UserModelExAccToken } from './userModelExAccToken';
import type { UserModelExAccTokenExp } from './userModelExAccTokenExp';
import type { UserModelGgAccToken } from './userModelGgAccToken';
import type { UserModelGgId } from './userModelGgId';
import type { UserModelGgReToken } from './userModelGgReToken';
import type { UserModelGhAccToken } from './userModelGhAccToken';
import type { UserModelGhId } from './userModelGhId';
import type { UserModelGhReToken } from './userModelGhReToken';
import type { UserModelProfileImgUri } from './userModelProfileImgUri';
import type { RoleType } from './roleType';
import type { UserModelTwofaSecret } from './userModelTwofaSecret';
import type { UserModelUserId } from './userModelUserId';

export interface UserModel {
  birthday?: UserModelBirthday;
  created_at?: string;
  /** @maxLength 64 */
  email: string;
  ex_acc_token?: UserModelExAccToken;
  ex_acc_token_exp?: UserModelExAccTokenExp;
  /** @maxLength 512 */
  full_name: string;
  gg_acc_token?: UserModelGgAccToken;
  gg_id?: UserModelGgId;
  gg_re_token?: UserModelGgReToken;
  gh_acc_token?: UserModelGhAccToken;
  gh_id?: UserModelGhId;
  gh_re_token?: UserModelGhReToken;
  is_2fa_enabled?: boolean;
  /** @maxLength 1024 */
  password_hash: string;
  profile_img_uri?: UserModelProfileImgUri;
  role?: RoleType;
  twofa_secret?: UserModelTwofaSecret;
  updated_at?: string;
  user_id?: UserModelUserId;
  /** @maxLength 32 */
  username: string;
}